requirements.txt
django==3.2.11 
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
web3==5.12.2
django-ipware==4.0.2
Pillow==9.0.0

